+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/global-121/121-platform/compare/v1.105.1...master)
+## [Unreleased](https://github.com/global-121/121-platform/compare/v1.105.7...master)
 
 ## Removed
 
@@ -13,6 +13,38 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   See: A step in the "How to set up a new instance"-guide in the wiki.
 
 ---
+
+## [1.105.7](https://github.com/global-121/121-platform/compare/v1.105.6...v1.105.7) - 2023-11-15
+
+### Fixed
+
+- Cannot include completed PA with 0 remaining payments (additional fix)
+
+## [1.105.6](https://github.com/global-121/121-platform/compare/v1.105.5...v1.105.6) - 2023-11-15
+
+### Fixed
+
+- Cannot include completed PA with 0 remaining payments
+
+## [1.105.5](https://github.com/global-121/121-platform/compare/v1.105.4...v1.105.5) - 2023-11-15
+
+### Fixed
+
+- paymentCount in PA popup
+- Setting of completed status & paymentCount
+
+## [1.105.4](https://github.com/global-121/121-platform/compare/v1.105.3...v1.105.4) - 2023-11-13
+
+### Fixed
+
+- Don't update last message status (performance issue) (additional places)
+
+## [1.105.3](https://github.com/global-121/121-platform/compare/v1.105.2...v1.105.3) - 2023-11-13
+
+### Fixed
+
+- Don't update last message status (performance issue)
+
 ## [1.105.2](https://github.com/global-121/121-platform/compare/v1.105.1...v1.105.2) - 2023-11-07
 
 ### Fixed